fixing all failing test cases

+++ db_test.go
@@ -2268,7 +2268,7 @@ func TestOpenDBReadOnly(t *testing.T) {
 	require.NoError(t, err)
 	// Add bunch of entries that go into value log.
 	require.NoError(t, db.Update(func(txn *Txn) error {
-		require.Greater(t, db.opt.ValueThreshold, 10)
+		require.Greater(t, db.opt.ValueThreshold, int64(10))
 		val := make([]byte, db.opt.ValueThreshold+10)
 		rand.Read(val)
 		for i := 0; i < 10; i++ {
